Backendpy/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
mysql-connector-python==8.4.0
scikit-learn==1.5.0
numpy==1.26.4
pydantic==2.7.1 